+++ Runtime/Scripts/Manager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 00f330f107d68bd3787ac903dfdbeb8d
+guid: 2a36be0ffd122e1f5a43d44acd87b3e0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
